Refactor ranking utilities into rank module
